+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Unreleased
-
+ - 2016-08-03
+   - Feature: Analysis Jobs items integration. Analysis jobs have been setup and their complete workflows tested and 
+     integrated. See [#300](https://github.com/QutBioacoustics/baw-server/pull/300)
  - 2016-06-11
    - Feature: Analysis Jobs item tracking - basic support added for tracking each item in an analysis job [#290](https://github.com/QutBioacoustics/baw-server/pull/290)
    - Enhancement: Filter::Query can now have it's base query customised [931066b](https://github.com/QutBioacoustics/baw-server/commit/931066b5eb30925e34e659bdff91e5b55abce764)